Add feature add player to admin tools on team page. Fix error when deleting a team that has players. Modify deleteTeam method to deleting all players and matches by team before deleting the team.

+++ src/main/java/fi/ollimyy/buttonleague/web/TeamController.java
@@ -1,9 +1,8 @@
 package fi.ollimyy.buttonleague.web;
 
-import fi.ollimyy.buttonleague.domain.PlayerRepository;
-import fi.ollimyy.buttonleague.domain.Team;
-import fi.ollimyy.buttonleague.domain.TeamRepository;
+import fi.ollimyy.buttonleague.domain.*;
 import fi.ollimyy.buttonleague.model.TeamStats;
+import fi.ollimyy.buttonleague.service.PlayerService;
 import fi.ollimyy.buttonleague.service.TeamStatsService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 public class TeamController {
@@ -26,7 +26,13 @@ public class TeamController {
     PlayerRepository playerRepository;
 
     @Autowired
+    MatchRepository matchRepository;
+
+    @Autowired
     TeamStatsService teamStatsService;
+
+    @Autowired
+    PlayerService playerService;
 
     @GetMapping("/league-table")
     public String showLeagueTable(Model model) {
@@ -44,13 +50,31 @@ public class TeamController {
     }
 
     @GetMapping("/team/{teamId}")
-    public String listAllPlayersByTeam(@PathVariable Long teamId, Model model, HttpSession session) {
-        model.addAttribute("team", teamRepository.findById(teamId).get()); //TODO: handle team not found
-        model.addAttribute("players", playerRepository.findPlayersByTeamId(teamId));
-        model.addAttribute("errorMessage", session.getAttribute("errorMessage"));
-        session.removeAttribute("errorMessage");
+    public String showTeamPage(@PathVariable Long teamId, Model model, HttpSession session) {
 
-        return "team-page";
+        Optional<Team> teamOptional = teamRepository.findById(teamId);
+        if (teamOptional.isEmpty()) {
+            session.setAttribute("errorMessage", "Team not found.");
+            return "redirect:/team-list";
+        } else {
+            Team team = teamOptional.get();
+
+
+            // for showing the team's name and it's players
+            model.addAttribute("team", team);
+            model.addAttribute("players", playerRepository.findPlayersByTeamId(teamId));
+
+            // for adding a new player to the team
+            Player newPlayer = new Player();
+            newPlayer.setTeam(team);
+            model.addAttribute("newPlayer", newPlayer); //for adding new players to the team
+            model.addAttribute("availableNumbers", playerService.getAvailablePlayerNumbersInTeam(team));
+
+            model.addAttribute("errorMessage", session.getAttribute("errorMessage"));
+            session.removeAttribute("errorMessage"); // remove error message from session so when user refreshes page it won't show again
+
+            return "team-page";
+        }
     }
 
     @GetMapping("/team-list")
@@ -83,9 +107,24 @@ public class TeamController {
 
     @GetMapping("/delete-team/{teamId}")
     @Secured("ADMIN")
-    public String deleteTeam(@PathVariable("teamId") Long teamId) {
-        teamRepository.deleteById(teamId);
+    public String deleteTeam(@PathVariable("teamId") Long teamId, HttpSession session) {
 
+        Optional<Team> teamOptional = teamRepository.findById(teamId);
+        if (teamOptional.isEmpty()) {
+            session.setAttribute("errorMessage", "Team not found.");
+        } else {
+            Team team = teamOptional.get();
+
+            // Delete all players associated with the team
+            List<Player> players = playerRepository.findPlayersByTeamId(teamId);
+            playerRepository.deleteAll(players);
+
+            // Delete all matches where the team is either the home team or away team
+            List<Match> matches = matchRepository.findMatchesPlayedByTeam(team);
+            matchRepository.deleteAll(matches);
+
+            teamRepository.deleteById(teamId);
+        }
         return "redirect:/team-list";
     }
